refatorado e funcionando

+++ src/main/java/io/brunoonofre64/domain/dto/orderitems/OrderItemsInputDTO.java
@@ -4,8 +4,6 @@ import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.Setter;
 
-import java.math.BigDecimal;
-
 @Getter
 @Setter
 @AllArgsConstructor
@@ -13,5 +11,5 @@ public class OrderItemsInputDTO {
 
     private String uuidProduct;
 
-    private BigDecimal amount;
+    private Long amount;
 }
